+++ src/backend/mailbox_manager/tests/api/mailboxes/test_api_mailboxes_create.py
@@ -496,7 +496,8 @@ def test_api_mailboxes__handling_dimail_unexpected_error():
 @mock.patch.object(Logger, "info")
 def test_api_mailboxes__send_correct_logger_infos(mock_info, mock_error):
     """
-    Upon requesting mailbox creation, things are correctly logged
+    Upon requesting mailbox creation, la régie should impersonate
+    querying user in dimail and log things correctly.
     """
     access = factories.MailDomainAccessFactory(role=enums.MailDomainRoleChoices.OWNER)
 
@@ -536,6 +537,9 @@ def test_api_mailboxes__send_correct_logger_infos(mock_info, mock_error):
         )
         assert response.status_code == status.HTTP_201_CREATED
 
+        # user sub is sent to payload as a parameter
+        assert rsps.calls[0].request.params == {"username": access.user.sub}
+
     # Logger
     assert not mock_error.called
     assert mock_info.call_count == 3
@@ -543,6 +547,7 @@ def test_api_mailboxes__send_correct_logger_infos(mock_info, mock_error):
         "Token succesfully granted by mail-provisioning API.",
     )
     assert mock_info.call_args_list[1][0] == (
-        "Mailbox successfully created on domain %s",
-        access.domain.name,
+        "Mailbox successfully created on domain %s by user %s",
+        str(access.domain),
+        access.user.sub,
     )