+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to
 
 - ✨(domains) add endpoint to list and retrieve domain accesses #404
 - 🍱(dev) embark dimail-api as container #366
-
+- ✨(dimail) allow la regie to request a token for another user #416 
 
 ### Changed